% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{call_uses_name}
\alias{call_uses_name}
\title{Checks if a call references a name.}
\usage{
call_uses_name(call, name)
}
\arguments{
\item{call}{A call.}

\item{name}{A name.}
}
\description{
Checks if a call references a name.
}
