% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.SPQR.R
\name{coef.SPQR}
\alias{coef.SPQR}
\title{coef method for class `SPQR`}
\usage{
\method{coef}{SPQR}(object, newx, ...)
}
\arguments{
\item{object}{an object of class \code{SPQR}}

\item{newx}{covariate values for which the coefficient is calculated}
}
\description{
Computes the estimated spline coefficients of a `SPQR` class object
}
