% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolds.SPQR.R
\name{createFolds.SPQR}
\alias{createFolds.SPQR}
\title{generate cross-validation folds}
\usage{
createFolds.SPQR(Y, nfold, stratified = FALSE)
}
\arguments{
\item{Y}{The response vector.}

\item{nfold}{The number of cross-validation folds.}

\item{stratified}{If \code{TRUE}, stratified folds based on quantiles of `Y` are generated.}
}
\value{
A list of size \code{nfold} containing indices of the observations for each fold.
}
\description{
Helper function to generate cross-validation folds that can be used by \code{cv.SPQR}.
}
