% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{lplot}
\alias{lplot}
\title{Plot with \code{type = "l", bty = "l", lwd = 1.5} as standard arguments.}
\usage{
lplot(...)
}
\arguments{
\item{...}{Arguments passed to \code{plot}}
}
\description{
Plot with \code{type = "l", bty = "l", lwd = 1.5} as standard arguments.
}
