% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcTrace.R
\name{mcmcTrace}
\alias{mcmcTrace}
\title{plot MCMC trace plots}
\usage{
mcmcTrace(
  object,
  target = c("loglik", "PDF", "CDF", "QF"),
  X = NULL,
  Y = NULL,
  tau = 0.5,
  window = NULL
)
}
\description{
Show trace plot of the log-likelihood or estimates of a `SPQR` class object fitted using the MCMC method
}
