% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_function.R
\name{P}
\alias{P}
\title{Create a function from an partial function call}
\usage{
P(call, .quote = TRUE)
}
\arguments{
\item{.quote}{Logical; If \code{TRUE}, the call is quoted.}

\item{.call}{A partial function call.}
}
\value{
A function.
}
\description{
Partial function function application is usually done by using the
\code{function} keyword to define a new function with some arguments
pre-specified. This function allows you to make a partially applied function
directly from a partial function call.
}
