% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SPQR.R
\name{plot.SPQR}
\alias{plot.SPQR}
\title{plot method for SPQR}
\usage{
\method{plot}{SPQR}(
  object,
  X,
  nY = 501,
  type = c("QF", "PDF", "CDF"),
  tau = seq(0.05, 0.95, 0.05),
  ci.level = 0,
  getAll = FALSE
)
}
\description{
Computes and plots the estimated PDF/CDF/QF curves
}
