% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQVI.R
\name{plotQVI}
\alias{plotQVI}
\title{Plot quantile ALE-induced variable importance}
\usage{
plotQVI(
  object,
  var.index = NULL,
  var.names = NULL,
  tau = 0.5,
  ci.level = 0,
  n.bins = 40
)
}
\arguments{
\item{object}{An object of class \code{"SPQR"}.}

\item{var.index}{A vector specifying the index of the covariates for which VI measures should be computed.
Default is \code{NULL} indicating all covariates are considered.}

\item{var.names}{The names of the covariates to appear in the bar plots. Default is \code{NULL} and the
the function will use generic names generated by \code{parse(text=paste0("X[",var.index,"]"))}.}

\item{tau}{The quantiles of interest.}

\item{ci.level}{The credible level for plotting the error bars. The default is 0 indicating no error bars should be plotted.}

\item{n.bins}{the maximum number of intervals into which the covariate range is divided when calculating the ALEs.
See \code{QALE()} for more detail.}
}
\value{
A \code{ggplot} object.
}
\description{
Computes the quantile ALE-induced variable importance (VI) measure for each of the
covariate specified in var.index, and produces a ranking plot of the covariates using
bar plot for each quantile of interest.
}
