% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.SPQR.R
\name{predict.SPQR}
\alias{predict.SPQR}
\title{predict method for class `SPQR`}
\usage{
\method{predict}{SPQR}(
  object,
  X,
  Y = NULL,
  nY = 501,
  type = c("QF", "PDF", "CDF"),
  tau = seq(0.05, 0.95, 0.05),
  ci.level = 0,
  getAll = FALSE,
  ...
)
}
\description{

}
