% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.SPQR.R
\name{print.summary.SPQR}
\alias{print.summary.SPQR}
\title{print method for summary.SPQR}
\usage{
\method{print}{summary.SPQR}(x, showModel = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{summary.SPQR}}

\item{showModel}{If \code{TRUE}, prints the detailed NN architecture by layer.}

\item{...}{Other arguments.}
}
\description{
Print the output produced by summary.SPQR().
}
