% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqCheck.R
\name{qqCheck}
\alias{qqCheck}
\title{Goodness-of-fit test for SPQR estimator}
\usage{
qqCheck(object, ci.level = 0, getAll = FALSE)
}
\description{
Performs a goodness-of-fit test for the estimated probability density function (PDF) using
inverse transformation method.
}
