% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_function.R
\name{R}
\alias{R}
\title{Allow recursive argmuments in function calls}
\usage{
R(.call, .quote = TRUE)
}
\arguments{
\item{.call}{A call with potentially recursive arguments.}

\item{.quote}{Logical; Is the call quoted?}
}
\value{
An evaluated call.
}
\description{
Allow recursive argmuments in function calls
}
