% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQR.ADAM.R
\name{SPQR.ADAM}
\alias{SPQR.ADAM}
\title{Fitting SPQR by MLE or MAP method}
\usage{
SPQR.ADAM(
  X,
  Y,
  n.knots,
  n.hidden,
  activation,
  method,
  prior,
  hyperpar,
  control,
  verbose = TRUE,
  seed = NULL
)
}
\description{
Fitting SPQR by MLE or MAP method
}
