% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQR.MCMC.R
\name{SPQR.MCMC}
\alias{SPQR.MCMC}
\title{Fitting SPQR by MCMC method}
\usage{
SPQR.MCMC(
  X,
  Y,
  n.knots,
  n.hidden,
  activation,
  prior,
  hyperpar,
  control,
  verbose = TRUE,
  seed = seed
)
}
\description{
Fitting SPQR by MCMC method
}
