% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQR.R
\docType{package}
\name{SPQR}
\alias{SPQR}
\title{Fitting SPQR models}
\usage{
SPQR(
  X,
  Y,
  n.knots = 12,
  n.hidden = 10,
  activation = c("tanh", "relu", "sigmoid"),
  method = c("MLE", "MAP", "MCMC"),
  prior = c("GP", "ARD", "GSM"),
  hyperpar = list(),
  control = list(),
  normalize = FALSE,
  verbose = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{X}{the covariate matrix}

\item{Y}{the response vector}

\item{n.knots}{number of spline basis to use}

\item{n.hidden}{number of hidden neurons}

\item{activation}{hidden layer activation}

\item{method}{type of estimator to use}

\item{prior}{prior model for variance hyperparameters}

\item{hyperpar}{hyperparameter values for the variance hyperpriors}

\item{control}{a list of named parameters for finer control of the training process}

\item{normalize}{whether to normalize the response and covariates to unit interval}

\item{verbose}{whether to print progress during training}

\item{seed}{random number generation seed}
}
\description{
Main function of the package. Fits SPQR using the maximum likelihood estimation (MLE), maximum \emph{a posterior} (MAP) or
Markov chain Monte Carlo (MCMC) method. Returns an object of S3 class `SPQR`.
}
