% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SPQR.R
\name{summary.SPQR}
\alias{summary.SPQR}
\title{summaary method for SPQR}
\usage{
\method{summary}{SPQR}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"SPQR"}.}

\item{...}{Other arguments.}
}
\value{
An object of class \code{"summary.SPQR"}.
}
\description{
summarizes the output produced by \code{SPQR()} and structures them in a more organized way to be examined by the user.
}
