% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{variables}
\alias{variables}
\title{Extracts variables from a call}
\usage{
variables(call, .standardise = TRUE)
}
\arguments{
\item{call}{A call}

\item{.standardise}{Should the calls be standardized?}
}
\value{
A list containing all the variables referenced in the call.
}
\description{
Extracts variables from a call
}
