% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.SPQR.R
\name{coef.SPQR}
\alias{coef.SPQR}
\title{coef method for class `SPQR`}
\usage{
\method{coef}{SPQR}(object, X, ...)
}
\arguments{
\item{object}{An object of class \code{SPQR}.}

\item{X}{The covariate vector/matrix for which the coefficient is calculated.}

\item{...}{Other arguments.}
}
\value{
A \code{NROW(X)} by K matrix containing values of the estimated coefficient, where K is the number of basis functions.
}
\description{
Computes the estimated spline coefficients of a `SPQR` class object
}
