% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolds.SPQR.R
\name{createFolds.SPQR}
\alias{createFolds.SPQR}
\title{generate cross-validation folds}
\usage{
createFolds.SPQR(Y, nfold, stratified = FALSE)
}
\description{
Helper function to generate cross-validation folds that can be used by \code{cv.SPQR}.
}
