% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.SPQR.R
\name{cv.SPQR}
\alias{cv.SPQR}
\title{Cross-validation for SPQR}
\usage{
cv.SPQR(
  X,
  Y,
  folds = NULL,
  n.knots = 12,
  n.hidden = 10,
  activation = c("tanh", "relu", "sigmoid"),
  method = c("MLE", "MAP"),
  prior = c("GP", "ARD", "GSM"),
  hyperpar = list(),
  control = list(),
  normalize = FALSE,
  verbose = TRUE,
  seed = NULL,
  ...
)
}
\description{
Fits SPQR using either MLE or MAP method. Computes K-fold cross-validation error
based on pre-computed folds.
}
