% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQVI.R
\name{plotQVI}
\alias{plotQVI}
\title{plot quantile ALE-induced variable importance}
\usage{
plotQVI(
  object,
  var.index = NULL,
  var.names = NULL,
  tau = 0.5,
  ci.level = 0,
  n.bins = 40
)
}
\description{

}
