% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SPQR.R
\name{print.SPQR}
\alias{print.SPQR}
\title{print method for SPQR}
\usage{
\method{print}{SPQR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SPQR}}

\item{...}{
  Arguments passed on to \code{\link[=print.summary.SPQR]{print.summary.SPQR}}
  \describe{
    \item{\code{showModel}}{If \code{TRUE}, prints the detailed NN architecture by layer.}
  }}
}
\description{
Summarizes and print the output produced by SPQR() in an organized way.
}
\details{
This is equivalent to the function call \code{print.summary.SPQR(summary.SPQR(object), ...)}.
}
