% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SPQR.R
\name{print.SPQR}
\alias{print.SPQR}
\title{print method for SPQR}
\usage{
\method{print}{SPQR}(x, showModel = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{SPQR}}

\item{showModel}{whether to print the NN structure}

\item{...}{other arguments}
}
\description{
print method for SPQR
}
