% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QALE.R
\name{QALE}
\alias{QALE}
\title{Quantile accumulated local effects (ALE)}
\usage{
QALE(
  object,
  var.index,
  tau,
  n.bins = 40,
  ci.level = 0,
  getAll = FALSE,
  pred.fun = NULL
)
}
\description{
Computes the quantile ALEs of a `SPQR` class object
}
