% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqCheck.R
\name{qqCheck}
\alias{qqCheck}
\title{Goodness-of-fit test for SPQR estimator}
\usage{
qqCheck(object, getAll = FALSE)
}
\arguments{
\item{object}{An object of class \code{"SPQR"}.}

\item{getAll}{If \code{TRUE} and \code{object$method == "MCMC"}, plots all posterior samples of Q-Q lines. Default: \code{FALSE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Performs a goodness-of-fit test for the estimated conditional probability density function (PDF) using
inverse transformation method.
}
