% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SPQR.R
\name{summary.SPQR}
\alias{summary.SPQR}
\title{summary method for SPQR}
\usage{
\method{summary}{SPQR}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"SPQR"}.}

\item{...}{Other arguments.}
}
\value{
An object of class \code{"summary.SPQR"}.
}
\description{
summarizes the output produced by \code{SPQR()} and structures them in a more organized way to be examined by the user.
}
\examples{
set.seed(919)
n <- 200
X <- rbinom(n, 1, 0.5)
Y <- rnorm(n, X, 0.8)
control <- list(iter = 300, warmup = 200, thin = 1)
fit <- SPQR(X = X, Y = Y, method = "MCMC", control = control, normalize = TRUE)

## summarize output
summary(fit)

}
