% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.SPQR.R, R/summary.SPQR.R
\name{summary.SPQR}
\alias{summary.SPQR}
\alias{print.summary.SPQR}
\title{summaary method for SPQR}
\usage{
\method{print}{summary.SPQR}(x, showModel = FALSE, ...)

\method{summary}{SPQR}(object, ...)
}
\arguments{
\item{x}{An object of class \code{SPQR}}

\item{showModel}{If \code{TRUE}, prints the detailed NN architecture by layer.}

\item{...}{Other arguments.}

\item{object}{An object of class \code{"SPQR"}.}
}
\value{
An object of class \code{"summary.SPQR"}.
}
\description{
summarizes the output produced by \code{SPQR()} and structures them in a more organized way to be examined by the user.
}
