% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.SPQR.R, R/summary.SPQR.R
\name{summary.SPQR}
\alias{summary.SPQR}
\alias{print.summary.SPQR}
\title{summaary method for SPQR}
\usage{
\method{print}{summary.SPQR}(x, showModel = FALSE, ...)

\method{summary}{SPQR}(object, ...)
}
\arguments{
\item{x}{an object of class \code{SPQR}}

\item{...}{other arguments}

\item{showMode}{whether to print the NN structure}
}
\description{
summaary method for SPQR
}
